Make JWT functionality optoinal with features
JWT support is not always needed. For instance, when writing something
like a GitHub Action, we can rely on PAT authentication and will never
need JWT.

JWT support currently requires `jsonwebtoken` v7, which currently has
several older dependencies (including base64, ring, etc). These
dependencies are updated in `jsonwebtoken` v8 (currently in
beta-status). The v8 branch has been in development for nearly two
years.

This change makes these dependencies optional by adding two new
features: `jwt` and `app`. The `app` feature depends on `jwt` and is
enabled by default to avoid making a backwards-incompatible change. This
allows users to avoid these dependencies by using `default-features =
false`.

This change moves JWT-specific types out of `lib.rs` into a `jwt` module
so they can be feature-gated together, as much as possible.

This change also updates CI to run `clippy` with no default features and
with only the `jwt` feature to ensure that these configurations don't
break.
