+++ Cargo.toml
@@ -26,7 +26,7 @@ dirs = { version = "3.0", optional = true }
 futures = { version = "0.3", default-features = false }
 http = "0.2"
 hyperx = "1"
-jsonwebtoken = "7"
+jsonwebtoken = { version = "7", optional = true }
 log = "0.4"
 mime = "0.3"
 percent-encoding = "2"
@@ -37,9 +37,11 @@ serde_json = "1.0"
 url = "2"
 
 [features]
-default = ["default-tls"]
+default = ["app", "default-tls"]
 # enable native tls
 default-tls = ["reqwest/default-tls"]
+app = ["jwt"]
+jwt = ["jsonwebtoken"]
 # enable rustls
 rustls-tls = ["reqwest/rustls-tls"]
 # enable etag-based http_cache functionality
